renamed old modules with old_ prefix, and updated notebooks under examples
