+++ notebooks/examples/example_user_interface_notebook.py
@@ -6,7 +6,7 @@
 #       extension: .py
 #       format_name: percent
 #       format_version: '1.3'
-#       jupytext_version: 1.19.1
+#       jupytext_version: 1.16.4
 #   kernelspec:
 #     display_name: cupix
 #     language: python
@@ -24,8 +24,8 @@ import h5py as h5
 from lace.cosmo import cosmology
 from cupix.px_data.data_DESI_DR2 import DESI_DR2
 from cupix.likelihood.likelihood_parameter import LikelihoodParameter
-from cupix.likelihood.likelihood import Likelihood
-from cupix.likelihood.lya_theory import Theory
+from cupix.likelihood.old_likelihood import Likelihood
+from cupix.likelihood.old_theory import Theory
 
 # %% [markdown]
 # ### Step 1: Load some data
